Fix heap allocator overlap, VFS deadlock/parent links, MADT remap, and PCI ECAM
- heap: Rework block splitting to never leave allocated blocks on the
  free list; use explicit back-pointer in `from_payload` instead of
  alignment rounding so dealloc works regardless of alignment padding.
- vfs: Set dentry parent links in `open`, `mkdir`, and path `walk_from`
  so relative paths and `..` resolution work correctly.
- vfs: Fix same-directory rename deadlock by locking the shared inode
  once instead of separately (non-reentrant spin mutex).
- acpi/madt: Add `phys_addr` parameter and fallback re-mapping when the
  higher-half vaddr does not contain a valid MADT signature. Add debug
  serial output for MADT entry parsing.
- acpi/tables: Carry `phys_addr` in `SdtEntry` for use by madt parser.
- pci/ecam: Fix bus number arithmetic to avoid u64 overflow on subtractions.

+++ kernel/src/pci/ecam.rs
@@ -69,7 +69,7 @@ static MAPPED: Mutex<Option<Vec<MappedRegion>>> = Mutex::new(None);
 pub fn map_all(regions: &PciConfigRegions) {
     let mut mapped = Vec::new();
     for entry in &regions.regions {
-        let num_buses = (entry.bus_number_end - entry.bus_number_start + 1) as u64;
+        let num_buses = entry.bus_number_end as u64 - entry.bus_number_start as u64 + 1;
         let size = num_buses << 20;
         let vaddr = map_ecam(entry.base_address, size);
         mapped.push(MappedRegion {
